Make subprocess output separate, invalidate QI/QCT
Open a separate output window for subprocess messages to not
spam the extension output window.
Make quantifier information disappear when verifying again (but
not when skipping) and when changing the document.
Also, do not consume the output of a to-be-killed verification
that is still taking place by checking the this.runCount variable.