Provide way to clear files affected by compilation
It is still a bit awkward, since now the infoCollection also has
to call the verificationManager. Maybe this should change.

Moreover, now the verification arguments are wrapped in a class
to allow a cleaner passing on of them in case they are needed in
a deeper method e.g. when maybe implementing a cleaner currentRun
check or when it is needed to know whether to actually clean
something provided in an earlier run.

Also, the setting of the prusti semantic version was moved to
also affect the server correctly.

And lastly, the thresholding mechanism of the diagnostics was
changed to be more predictable.